refactor(hopr-strategy): move channel_lifecycle tests to pipeline.rs and add restart test
- Move all unit tests from channel_lifecycle/mod.rs to pipeline.rs, co-located with the logic under test
- Add inflight_events_arrive_at_new_instance_after_restart: simulates a strategy instance being dropped and recreated while in-flight events from the old instance still arrive — removals are no-ops on the fresh instance, but ChannelClosed still applies the peer cooldown
- Switch all async channel_lifecycle tests to plain #[tokio::test] (no test_log wrapper)
